requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
requests==2.31.0
statsmodels==0.14.0
matplotlib==3.7.2
seaborn==0.12.2
python-dotenv==1.0.0